Further factoring out to use the task class throughout; test_forecast.py passes except for numpy and prophet
