+++ flaml/automl/state.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 
 from flaml import tune
-from flaml.data import TS_FORECAST
+from .tasks import TS_FORECAST
 from flaml.ml import compute_estimator, train_estimator
 from .ts_data import TimeSeriesDataset
