+++ flaml/default/estimator.py
@@ -1,6 +1,6 @@
 import sklearn.ensemble as ensemble
 from functools import wraps
-from flaml.data import CLASSIFICATION
+from ..automl.tasks import CLASSIFICATION
 from .suggest import preprocess_and_suggest_hyperparams
 
 DEFAULT_LOCATION = "default_location"
