+++ flaml/default/suggest.py
@@ -3,8 +3,9 @@ from sklearn.neighbors import NearestNeighbors
 import logging
 import pathlib
 import json
-from flaml.data import CLASSIFICATION, DataTransformer
-from flaml.ml import get_estimator_class, get_classification_objective
+from flaml.data import DataTransformer
+from flaml.automl.tasks import CLASSIFICATION, get_classification_objective
+from flaml.ml import get_estimator_class
 
 LOCATION = pathlib.Path(__file__).parent.resolve()
 logger = logging.getLogger(__name__)
