+++ test/automl/test_multiclass.py
@@ -3,7 +3,8 @@ import numpy as np
 import scipy.sparse
 from sklearn.datasets import load_iris, load_wine
 from flaml import AutoML
-from flaml.data import CLASSIFICATION, get_output_from_log
+from flaml.data import get_output_from_log
+from flaml.automl.tasks import CLASSIFICATION
 from flaml.model import LGBMEstimator, XGBoostSklearnEstimator, SKLearnEstimator
 from flaml import tune
 from flaml.training_log import training_log_reader
